fix(connection): fix connection ResolveConnFileDesc func
When a TCP/UDP connection returns a file, it is a copy of the underlying OS file, and closing this file with syscall.close(file.FD()) has no effect on closing the connection.

Thus, it is necessary to use the SyscallConn method to obtain the actual connection and invoke Control(f func(fd uintptr)) error to obtain the actual FD for effective connection management.